backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
httpx>=0.28.0
beautifulsoup4>=4.12.0
lxml>=5.3.0
networkx>=3.4.0
google-auth>=2.37.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.20
aiofiles>=24.1.0
openai>=1.59.0
anthropic>=0.42.0
sqlalchemy>=2.0.0
